% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_for_smxapp.R
\name{munge_for_smxapp}
\alias{munge_for_smxapp}
\title{Create R-binary bundle for the smxapp}
\usage{
munge_for_smxapp(con, st, nu, le, kv, id = 30, gid = 73, cruise)
}
\arguments{
\item{con}{XXX}

\item{st}{a list contain st, nu, le and kv or a st dataframe}

\item{nu}{nu dataframe, not used if st is a list}

\item{le}{le dataframe, not used if st is a list}

\item{kv}{kv dataframe, not used if st is a list}

\item{id}{synaflokkur}

\item{gid}{veidarfaeri}

\item{cruise}{leidangrar í schema hafvog.}
}
\description{
Create R-binary bundle for the smxapp
}
